rename `self-build` to `build-self`
