+++ xtask/src/build_self.rs
